\name{div_utilities}
\alias{col_names}
\alias{likelihood_func}
\alias{simdiv_outsource}
\alias{div_utilities}
\title{
Internal Utility Functions
}
\description{
Utility functions not directly called by the user. 
}
\details{
Three utility functions are called upon internally by 'diverge' and typically will not be called by the user. These include simdiv_outsource, col_names, and likelihood_func. Typing 'likelihood_func' may however be of interest to those wishing see/verify how we coded up the likelihood functions for each evolutionary model.
}